+++ packages/moonstone/Slider/Slider.js
@@ -22,13 +22,16 @@ import {SliderBarFactory} from './SliderBar';
 import SliderTooltip from './SliderTooltip';
 import componentCss from './Slider.less';
 
-const isActive = (ev, props) => props.active || props.detachedKnob;
+const isActive = (ev, props) => props.active || props.directControl || props.detachedKnob;
 const isIncrement = (ev, props) => forKey(props.vertical ? 'up' : 'right', ev);
 const isDecrement = (ev, props) => forKey(props.vertical ? 'down' : 'left', ev);
+// If directControl is enabled -> and not active, run onActivate, otherwise (it's already active) just return true and continue
+const isDirectlyControlled = (ev, props) => (props.directControl && (props.active || forward('onActivate', ev, props)));
 
 const handleDecrement = handle(
 	isActive,
 	isDecrement,
+	isDirectlyControlled,
 	forward('onDecrement'),
 	stopImmediate
 );
@@ -36,6 +39,7 @@ const handleDecrement = handle(
 const handleIncrement = handle(
 	isActive,
 	isIncrement,
+	isDirectlyControlled,
 	forward('onIncrement'),
 	stopImmediate
 );
@@ -109,6 +113,14 @@ const SliderBaseFactory = factory({css: componentCss}, ({css}) => {
 			 * @public
 			 */
 			detachedKnob: PropTypes.bool,
+
+			/**
+			 * When `true`, the component may be manipulated via the directional input keys
+			 *
+			 * @type {Boolean}
+			 * @public
+			 */
+			directControl: PropTypes.bool,
 
 			/**
 			 * When `true`, the component is shown as disabled and does not generate events
@@ -336,6 +348,7 @@ const SliderBaseFactory = factory({css: componentCss}, ({css}) => {
 			backgroundProgress: 0,
 			knobAfterMidpoint: false,
 			detachedKnob: false,
+			directControl: false,
 			focused: false,
 			max: 100,
 			min: 0,
@@ -402,6 +415,7 @@ const SliderBaseFactory = factory({css: componentCss}, ({css}) => {
 		render: ({backgroundProgress, children, disabled, focused, inputRef, knobAfterMidpoint, max, min, onBlur, onChange, onKeyDown, onMouseMove, onMouseUp, proportionProgress, scrubbing, sliderBarRef, sliderRef, step, tooltip, tooltipForceSide, tooltipSide, value, vertical, ...rest}) => {
 			delete rest.active;
 			delete rest.detachedKnob;
+			delete rest.directControl;
 			delete rest.noFill;
 			delete rest.onActivate;
 			delete rest.onDecrement;